requirements.txt
pvporcupine==3.0.2
openai-whisper==20231117
insightface==0.7.3
opencv-python==4.8.1.78
numpy==1.26.2
pyaudio==0.2.13
pyttsx3==2.90
requests==2.31.0
onnxruntime==1.16.3
pillow==10.1.0
flet==0.21.2
edge-tts
httpx
pynput
rumps
chromadb
